Fix some warning and a minor template compatibility issue

+++ gwen/src/Platforms/Windows.cpp
@@ -144,7 +144,7 @@ bool Gwen::Platform::FileOpen( const String& Name, const String& StartPath, cons
 	char FilterBuffer[512];
 	{
 		memset( FilterBuffer, 0, sizeof(FilterBuffer) );
-		memcpy( FilterBuffer, Extension.c_str(), Gwen::Min( Extension.length(), 512 ) );
+		memcpy( FilterBuffer, Extension.c_str(), Gwen::Min( (int)Extension.length(), 512 ) );
 		for (int i=0; i<512; i++)
 		{
 			if ( FilterBuffer[i] == '|' )
@@ -256,7 +256,7 @@ bool Gwen::Platform::FileSave( const String& Name, const String& StartPath, cons
 	char FilterBuffer[512];
 	{
 		memset( FilterBuffer, 0, sizeof(FilterBuffer) );
-		memcpy( FilterBuffer, Extension.c_str(), Gwen::Min( Extension.size(), 512 ) );
+		memcpy( FilterBuffer, Extension.c_str(), Gwen::Min( (int)Extension.size(), 512 ) );
 		for (int i=0; i<512; i++)
 		{
 			if ( FilterBuffer[i] == '|' )
